+++ src/engraving/dom/dom.cmake
@@ -215,6 +215,8 @@ set(DOM_SRC
     ${CMAKE_CURRENT_LIST_DIR}/ottava.h
     ${CMAKE_CURRENT_LIST_DIR}/page.cpp
     ${CMAKE_CURRENT_LIST_DIR}/page.h
+    ${CMAKE_CURRENT_LIST_DIR}/pagelockindicator.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pagelockindicator.h
     ${CMAKE_CURRENT_LIST_DIR}/parenthesis.cpp
     ${CMAKE_CURRENT_LIST_DIR}/parenthesis.h
     ${CMAKE_CURRENT_LIST_DIR}/partialtie.cpp
@@ -322,6 +324,8 @@ set(DOM_SRC
     ${CMAKE_CURRENT_LIST_DIR}/system.h
     ${CMAKE_CURRENT_LIST_DIR}/systemdivider.cpp
     ${CMAKE_CURRENT_LIST_DIR}/systemdivider.h
+    ${CMAKE_CURRENT_LIST_DIR}/systemlockindicator.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/systemlockindicator.h
     ${CMAKE_CURRENT_LIST_DIR}/systemtext.cpp
     ${CMAKE_CURRENT_LIST_DIR}/systemtext.h
     ${CMAKE_CURRENT_LIST_DIR}/tapping.cpp
