+++ src/engraving/dom/select.h
@@ -132,7 +132,7 @@ public:
     MeasureBase* startMeasureBase() const;
     MeasureBase* endMeasureBase() const;
     std::vector<System*> selectedSystems() const;
-    std::vector<Page*> selectedPages() const;
+    std::vector<Page*> pagesContainingSelection() const;
     void update();
     void updateState();
     void dump();
