+++ src/engraving/dom/systemlockindicator.cpp
@@ -1,0 +1,52 @@
+/*
+ * SPDX-License-Identifier: GPL-3.0-only
+ * MuseScore-Studio-CLA-applies
+ *
+ * MuseScore Studio
+ * Music Composition & Notation
+ *
+ * Copyright (C) 2026 MuseScore Limited and others
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#include "systemlockindicator.h"
+
+#include "measure.h"
+#include "measurebase.h"
+#include "page.h"
+#include "rangelock.h"
+#include "system.h"
+
+namespace mu::engraving {
+SystemLockIndicator::SystemLockIndicator(System* parent, const RangeLock* lock)
+    : IndicatorIcon(ElementType::SYSTEM_LOCK_INDICATOR, parent, ElementFlag::SYSTEM | ElementFlag::GENERATED), m_systemLock(lock) {}
+
+void SystemLockIndicator::setSelected(bool v)
+{
+    EngravingItem::setSelected(v);
+    renderer()->layoutItem(this);
+    system()->page()->invalidateBspTree();
+}
+
+String SystemLockIndicator::formatBarsAndBeats() const
+{
+    const MeasureBase* startMB = m_systemLock->startMB();
+    const MeasureBase* endMB = m_systemLock->endMB();
+    const Measure* startMeasure = startMB->isMeasure() ? toMeasure(startMB) : toMeasure(startMB->prevMeasure());
+    const Measure* endMeasure = endMB->isMeasure() ? toMeasure(endMB) : toMeasure(endMB->prevMeasure());
+    const int startMeasureNum = startMeasure ? startMeasure->measureNumber() : -1;
+    const int endMeasureNum = endMeasure ? endMeasure->measureNumber() : -1;
+    return muse::mtrc("engraving", "Start measure: %1; End measure: %2").arg(startMeasureNum).arg(endMeasureNum);
+}
+} // namespace mu::engraving
