+++ src/engraving/dom/utils.cpp
@@ -117,22 +117,6 @@ Measure* Score::tick2measureMM(const Fraction& t) const
     return measure->coveringMMRestOrThis();
 }
 
-//---------------------------------------------------------
-//   measureAtTick
-//---------------------------------------------------------
-
-Measure* Score::measureAtTick(const Fraction& tick) const
-{
-    std::vector<MeasureBase*> mbList = m_measures.measureBasesAtTick(tick.ticks());
-    for (MeasureBase* mb : mbList) {
-        if (mb->isMeasure()) {
-            return toMeasure(mb);
-        }
-    }
-    LOGD("measureAtTick %d not found", tick.ticks());
-    return nullptr;
-}
-
 MeasureBase* Score::tick2measureBase(const Fraction& tick) const
 {
     if (tick == Fraction(-1, 1)) {   // special number
@@ -142,7 +126,7 @@ MeasureBase* Score::tick2measureBase(const Fraction& tick) const
         return m_measures.first();
     }
 
-    return m_measures.measureBaseByTick(tick.ticks());
+    return m_measures.firstMeasureBaseAtTick(tick.ticks());
 }
 
 //---------------------------------------------------------
