+++ src/notation/internal/notationselection.h
@@ -56,7 +56,7 @@ public:
     mu::engraving::MeasureBase* startMeasureBase() const override;
     mu::engraving::MeasureBase* endMeasureBase() const override;
     std::vector<mu::engraving::System*> selectedSystems() const override;
-    std::vector<mu::engraving::Page*> selectedPages() const override;
+    std::vector<mu::engraving::Page*> pagesContainingSelection() const override;
 
     bool elementsSelected(const mu::engraving::ElementTypeSet& types) const override;
